Move classes to its own files
