+++ tools/Titan.Tools.ManifestBuilder/DataTemplates/Attributes/EditorEnumAttribute.cs
@@ -1,0 +1,8 @@
+using System;
+
+namespace Titan.Tools.ManifestBuilder.DataTemplates.Attributes;
+
+[AttributeUsage(AttributeTargets.Property)]
+public class EditorEnumAttribute : EditorNodeAttribute
+{
+}
