+++ tools/Titan.Tools.ManifestBuilder/DataTemplates/Attributes/EditorNodeAttribute.cs
@@ -1,0 +1,6 @@
+using System;
+
+namespace Titan.Tools.ManifestBuilder.DataTemplates.Attributes;
+
+[AttributeUsage(AttributeTargets.Property)]
+public abstract class EditorNodeAttribute : Attribute { }
