+++ tools/Titan.Tools.ManifestBuilder/DataTemplates/Attributes/EditorStringAttribute.cs
@@ -1,0 +1,9 @@
+using System;
+
+namespace Titan.Tools.ManifestBuilder.DataTemplates.Attributes;
+
+[AttributeUsage(AttributeTargets.Property)]
+public class EditorStringAttribute : EditorNodeAttribute
+{
+    public string? Watermark { get; set; }
+}
