+++ tools/Titan.Tools.ManifestBuilder/Models/ManifestItem.cs
@@ -1,0 +1,9 @@
+using Titan.Tools.ManifestBuilder.DataTemplates.Attributes;
+
+namespace Titan.Tools.ManifestBuilder.Models;
+
+public class ManifestItem
+{
+    [EditorString]
+    public required string Name { get; set; }
+}
