+++ tools/Titan.Tools.ManifestBuilder/Models/ModelItem.cs
@@ -1,0 +1,6 @@
+namespace Titan.Tools.ManifestBuilder.Models;
+
+public class ModelItem : ManifestItem
+{
+    public required string Path { get; set; }
+}
