+++ tools/Titan.Tools.ManifestBuilder/Models/TextureItem.cs
@@ -1,0 +1,10 @@
+using System.Text.Json.Serialization;
+
+namespace Titan.Tools.ManifestBuilder.Models;
+
+public class TextureItem : ManifestItem
+{
+    public required string Path { get; init; }
+    [JsonConverter(typeof(JsonStringEnumConverter))]
+    public TextureType Type { get; set; }
+}
