+++ tools/Titan.Tools.ManifestBuilder/Models/TextureType.cs
@@ -1,0 +1,9 @@
+namespace Titan.Tools.ManifestBuilder.Models;
+
+public enum TextureType
+{
+    Unknown,
+    PNG,
+    BMP,
+    JPG
+}
